Augment test environment to support simul I$/D$ sequences
The instances being targeted for overriding must have a set
context whose name will be pattern matched against the instance name
given in ::type_id::set_inst_override.  In this case, the agent path was targeted.

The context of uvm_sequence's appears to be set when uvm_sequence_base.start()
is called, but it is manually provided during ::type_id::create() regardless.
However, uvm_sequence_item's must be given a context in order to be
overridden.  The context of the sequencer (accessed with
get_full_name()) is used because the type override in l1_basic_test
targets the agent as a whole.

The driver will output the correct uvm_sequence_item type because it was
given by the targeted uvm_sequence, but the monitor is unaffected.  As a
result, the scoreboard will raise a type conflict error.  To fix this,
the agent as a whole is targeted for instance overriding in
::type_id::set_inst_override.  The agent as a whole now experiences
the uvm_sequence_item type overriding.

Kudos to these forums:
https://verificationacademy.com/forums/t/uvm-instance-override-for-uvm-objects/38640
https://verificationacademy.com/forums/t/overriding-transaction-item-in-specific-sequence-instance/30142